Fixes need to work with ARDrone 2.0 including deal with vidoe streaming starting with an incomplete header (as seems to happen with my drone).
- Added tests and travis.ci support.
