+++ dev-requirements.txt
@@ -1,0 +1,2 @@
+pytest==2.3.5
+pytest-xdist==1.8
