+++ requirements.txt
@@ -1,0 +1,3 @@
+PIL==1.1.7
+mock==1.0.1
+numpy==1.7.1
